src/scripts/requirements.txt
# Voice Note Transcription Dependencies
# Install with: pip install -r requirements.txt

openai-whisper>=20231117
torch>=2.0.0
tqdm>=4.65.0

# Optional: For faster GPU processing (NVIDIA only)
# Uncomment if you have a CUDA-capable GPU:
# --extra-index-url https://download.pytorch.org/whl/cu118
# torch>=2.0.0+cu118
